Appendix B - Source code/modules/LoggingController/requirements.txt
azure-iot-device~=2.7.0
numpy~=1.21.5